feat(dispatch): reorganize settings into settings/money groups
Pure UI reshuffle of the dispatch settings IA — no setting keys, scopes,
or migrations touched. Dispatch is unreleased, so routes rename freely.

- General is now org identity: business info + currency (from Documents),
  regional week-start/24h rows (from Availability), branding (from Documents)
- Availability renamed Scheduling; absorbs old General's Routes + Board
- New per-object money pages: Quotes (PDF defaults + acceptance/deposits),
  Invoicing (automation + invoice PDF defaults), Payments (Stripe + partial
  payments + receipt email); Documents page deleted, Preview PDF now on Quotes
- Sidebar gets Settings/Money header groups; getting-started links updated
- Each page owns its own useDirtyDraft/FormSaveBar slice
