--- apps/web/src/app/(protected)/app/dispatch/settings/availability/page.tsx
@@ -1,9 +1,0 @@
-// apps/web/src/app/(protected)/app/dispatch/settings/availability/page.tsx
-
-import { AvailabilitySettingsPage } from '~/components/availability/ui/availability-settings-page'
-
-function DispatchAvailabilitySettings() {
-  return <AvailabilitySettingsPage />
-}
-
-export default DispatchAvailabilitySettings
