--- apps/web/src/app/(protected)/app/dispatch/settings/documents/page.tsx
@@ -1,9 +1,0 @@
-// apps/web/src/app/(protected)/app/dispatch/settings/documents/page.tsx
-
-import { DocumentsSettingsPage } from '~/components/money/ui/settings/documents-page'
-
-function DispatchDocumentsSettings() {
-  return <DocumentsSettingsPage />
-}
-
-export default DispatchDocumentsSettings
