+++ apps/web/src/app/(protected)/app/dispatch/settings/quotes/page.tsx
@@ -1,0 +1,9 @@
+// apps/web/src/app/(protected)/app/dispatch/settings/quotes/page.tsx
+
+import { QuotesSettingsPage } from '~/components/money/ui/settings/quotes-page'
+
+function DispatchQuotesSettings() {
+  return <QuotesSettingsPage />
+}
+
+export default DispatchQuotesSettings
