+++ apps/web/src/app/(protected)/app/dispatch/settings/scheduling/page.tsx
@@ -1,0 +1,9 @@
+// apps/web/src/app/(protected)/app/dispatch/settings/scheduling/page.tsx
+
+import { SchedulingSettingsPage } from '~/components/availability/ui/scheduling-settings-page'
+
+function DispatchSchedulingSettings() {
+  return <SchedulingSettingsPage />
+}
+
+export default DispatchSchedulingSettings
